Fixed Bug 61102 - Web: Common: removed the window with deletion if th…

+++ packages/common/components/MediaViewer/index.tsx
@@ -273,6 +273,10 @@ function MediaViewer({
     let currentFileId = playlist.find((file) => file.id === playlistPos)
       ?.fileId;
 
+    const canDelete = targetFile?.security?.Delete;
+
+    if (!canDelete) return;
+
     if (!isNullOrUndefined(currentFileId)) onDelete(currentFileId);
   };
